Move LoggingCommandExecutor to tests.test_utils
